Add files via upload
As a grid operator, Alliander – the largest distribution system operator (DSO) in the Netherlands – required accurate forecasts of the load on the electricity grid looking hours to days ahead. With the rise in renewable energy and electrification of energy consumption pushing the capacity of grids even further, this need will only continue to grow. To address this important topic, Alliander started the project OpenSTEF (short for Short-Term-Energy-Forecasting) to anticipate congestion in the distribution grid, allow for grid safety analysis in the transmission grid, and enable smart grid innovations to locally balance supply and demand within the constraints of the grid. This is essential for anticipating local congestion levels and making the most of existing assets to achieve load balance and improve efficiency.

Signed-off-by: dbrownlf <120492450+dbrownlf@users.noreply.github.com>